# Use 2D block loads for post-DPAS chained ops 