This is a proof of concept to use the 2D block load for tensor pointer loads where the layout is a DPAS layout but the result of the load is not directly used in the DPAS computation. I built the PoC arond the `gemm_postop_addmatrix_benchmark` kernel which computes `AxB + D`. The `D` matrix uses a block pointer load and the TTGIR applies the DPAS MMA layout to the `D` matrix load:
```
    %16 = tt.make_tensor_ptr %arg3, [%c1024_i64, %c4096_i64], [%c4096_i64, %c1_i64], [%9, %12] {order = array<i32: 1, 0>} : <tensor<256x256xf32, #mma>> loc(#loc21)
    %17 = tt.load %16 {boundaryCheck = array<i32: 0, 1>, triton_intel_gpu.block_io = "row_major"} : !tt.ptr<tensor<256x256xf32, #mma>> loc(#loc22)
    %18 = arith.addf %15#0, %17 : tensor<256x256xf32, #mma> loc(#loc23)
    %19 = tt.make_tensor_ptr %arg2, [%c1024_i64, %c4096_i64], [%c4096_i64, %c1_i64], [%9, %12] {order = array<i32: 1, 0>} : <tensor<256x256xf32, #mma>> loc(#loc24)
    tt.store %19, %18 {boundaryCheck = array<i32: 0, 1>} : !tt.ptr<tensor<256x256xf32, #mma>> loc(#loc25)
```
But, when this code is lowered currently we use a scalar load. This requires extracting all the scalar values from the dpas registers, loading each scalar from a tensor of pointers for the `D` matrix, and computing the scalar add. A representative GEMM kernel for `AxB` is approximately 1100 instructions in ASM, but with scalar `D` matrix addition that increases to 7500 instructions and is quite slow. With this PR, the ASM is down to 1500 and performance is much better. I believe we both increase memory bandwidth by doing larger loads and decrease register pressure by using the 2D block load data directly from registers. I modified the GEMM with block pointers tutorial to do the matrix addition and compared directly with PyTorch.

main: 
```
Compute A x B
✅ Triton and Torch match
AxB      	torch			triton
gbps mean	156.7301576517312	138.11238212309073
gbps min	141.30597513931122	125.77424121785864
gbps max	165.34311994793367	142.42265041843044
tflops mean	97.2676857184077	85.71338139033023
tflops min	87.69534457130585	78.05625636793165
tflops max	102.61294231920246	88.38836001725622
```
w/ this change:
```
Compute A x B
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       157.91230857925294      157.63376432129215
gbps min        140.6169072268218       143.2243668186498
gbps max        167.52037117445636      163.5302792054239
tflops mean     98.00133574857878       97.82846949394128
tflops min      87.26770484864576       88.88591007411962
tflops max      103.96415762584442      101.48788236748732
```
note however that our GEMM performance is not as good as PyTorch/oneDNN:
```
Compute A x B
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       223.0701969650354       181.18122189537945
gbps min        187.16468723584453      162.91434700909443
gbps max        238.444109847901        192.58129503451758
tflops mean     138.43871617708862      112.44216437628396
tflops min      116.15553923000287      101.10563111352282
tflops max      147.97985968742458      119.51711885778543
```
But now with the matrix addition, we're even with / sometimes slightly ahead of PyTorch. This demonstrates the benefit of operator fusion in Triton and if we can improve gemm performance a bit we should be able to pull ahead of PyTorch for this use case. 

The code here contains a lot of duplication because I was not sure what parts of the existing block load code would be relevant, and what could be ignored. I plan to clean up the duplication (and I expect a few test failures that I will have to fix) before this is ready to be moved out of draft. 

cc #2997