@alexbaden 2024-12-12
Add matrix is noticeably improved (~5%), other benchmarks are inline w/ no regressions: https://benchmarks.glados.intel.com/d/1pXX4hUSz/microbenchmarks?orgId=1&var-tag=ci%7Cci-adv%7Cci-dflt%7Cpr3000&var-bench=All&var-device=Intel%28R%29%20Data%20Center%20GPU%20Max%201550&var-compiler=triton&var-compiler=xetla&var-compiler=onednn&var-backend=All&var-baseline_backend=xetla-ci-XPU%201550&var-target_backend=triton-ci-XPU%201550