@alexbaden 2024-12-12
This also resolves the performance regression in #2834 for the add matrix case. We get even to slightly-higher performance when we use a transposed `warpsPerCTA` order:
Standard warpsPerCTA Order:
```
Compute A x B
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       205.80269703051565      183.18315754413806
gbps min        179.1045897088753       162.24215503026394
gbps max        226.5780973097816       192.58129503451758
tflops mean     127.7224010662109       113.68457777284686
tflops min      111.15339385568986      100.68846469756984
tflops max      140.6157403910402       119.51711885778543
Compute A x B + D
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       157.65789768085992      158.24855403301189
gbps min        142.79881082578163      141.67625237645345
gbps max        168.10632973298533      163.16016456641276
tflops mean     97.84344680315183       98.2100117150328
tflops min      88.6218074458184        87.9251408687808
tflops max      104.32780705853148      101.25818697939796
```

Transposed `warpsPerCTA` order:
```
Compute A x B
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       220.02125371014108      189.8327155522294
gbps min        196.16217040160757      163.2217377764552
gbps max        234.8195487001742       196.96612401967587
tflops mean     136.54652351465722      117.81133377301992
tflops min      121.73943181287643      101.29639968672126
tflops max      145.7304350721081       122.23837030069579
Compute A x B + D
✅ Triton and Torch match
AxB             torch                   triton
gbps mean       161.31998225073477      166.80650419582096
gbps min        142.56348466801188      149.7706365256271
gbps max        168.8940264630101       176.18639363624158
tflops mean     100.1161586816681       103.52112745243674
tflops min      88.47576260608734       92.94856472863161
tflops max      104.81665642310443      109.34234368697658
```