@alexbaden 2024-12-13
I have been trying to break it today and have not uncovered any issues. I cleaned up a good chunk of the duplication and have some ideas for how to further abstract away the differences but I'm going to save that for another PR. This is now ready for review.