@alexbaden 2024-12-17
I modified the addmatrix benchmarks to include int8 to get some idea of correctness and performance there. The result is correct, and I see block loads in the IR. Performance is better:

```
block adds:
matmul-performance-postop-addmatrix:
   B     M     K     N           dtype  Triton-GB/s  Triton-GB/s-min  Triton-GB/s-max  Triton-TFlops  Triton-TFlops-min  Triton-TFlops-max  Triton-CV
0  1  1024  1024  1024  torch.bfloat16   209.401106       113.975650       210.557421      53.606683          29.177766          53.902700   0.002067
1  1  1024  1024  1024      torch.int8   236.698860       234.057137       239.400923      60.594908          59.918627          61.286636   0.006385
2  1  2048  2048  2048  torch.bfloat16   187.350255       160.332722       206.007068      95.923331          82.090354         105.475619   0.064107
3  1  2048  2048  2048      torch.int8   222.953037       219.367361       225.500214     114.151955         112.316089         115.456110   0.006912
4  1  4096  4096  4096  torch.bfloat16   164.567223       152.215714       168.344529     168.516837         155.868891         172.384797   0.029105
5  1  4096  4096  4096      torch.int8   164.680295       159.358059       169.227511     168.632622         163.182653         173.288971   0.013594
6  1  8192  8192  8192  torch.bfloat16   103.265063        95.490571       105.903398     211.486848         195.564689         216.890160   0.019848
7  1  8192  8192  8192      torch.int8    88.904751        87.769900        89.910054     182.076931         179.752756         184.135790   0.007747

main: 
matmul-performance-postop-addmatrix:
   B     M     K     N           dtype  Triton-GB/s  Triton-GB/s-min  Triton-GB/s-max  Triton-TFlops  Triton-TFlops-min  Triton-TFlops-max  Triton-CV
0  1  1024  1024  1024  torch.bfloat16   220.289058        76.875076       224.054711      56.393999          19.680019          57.358006   0.008403
1  1  1024  1024  1024      torch.int8   215.423907       213.995100       217.546892      55.148520          54.782745          55.692004   0.003764
2  1  2048  2048  2048  torch.bfloat16   175.145800       162.318271       182.361043      89.674650          83.106955          93.368854   0.021099
3  1  2048  2048  2048      torch.int8   186.579361       185.260784       187.748615      95.528633          94.853522          96.127291   0.001961
4  1  4096  4096  4096  torch.bfloat16   137.149996       121.170133       149.369803     140.441596         124.078216         152.954678   0.064511
5  1  4096  4096  4096      torch.int8   135.321954       123.090362       136.024125     138.569681         126.044530         139.288704   0.006765
6  1  8192  8192  8192  torch.bfloat16    88.742803        87.917075        89.290380     181.745260         180.054169         182.866699   0.002579
7  1  8192  8192  8192      torch.int8    71.957158        69.739437        73.489197     147.368260         142.826368         150.505874   0.010592
``` 

I will PR the benchmark changes separately - torch does not support int8 mma on the GPU so we have to go back to CPU, which makes the benchmark take about 20x longer. We will have to decide how to handle that.